Adding FlagRegistry to IAM Role
