+++ cloudaux/orchestration/aws/iam/role.py
@@ -2,6 +2,19 @@ from cloudaux import CloudAux
 from cloudaux.aws.iam import get_role_managed_policies, get_role_inline_policies, get_role_instance_profiles
 from cloudaux.orchestration.aws import _get_name_from_structure, _conn_from_args
 from cloudaux.orchestration import modify
+from cloudaux.orchestration.flag_registry import FlagRegistry as Registry
+from bunch import Bunch
+
+
+FLAGS=Bunch(
+    MANAGED_POLICIES=1,
+    INLINE_POLICIES=2,
+    INSTANCE_PROFILES=4,
+    ALL=7)
+
+class FlagRegistry(Registry):
+    from collections import defaultdict
+    r = defaultdict(list)
 
 
 def _get_base(role, **conn):
@@ -27,11 +40,27 @@ def _get_base(role, **conn):
 
     # cast CreateDate from a datetime to something JSON serializable.
     role.update(dict(CreateDate=str(role['CreateDate'])))
+    role['_version'] = 1
 
     return role
 
 
-def get_role(role, output='camelized', **conn):
+@FlagRegistry.register(flag=FLAGS.MANAGED_POLICIES, key='managed_policies')
+def get_managed_policies(role, **conn):
+    return get_role_managed_policies(role, **conn)
+
+
+@FlagRegistry.register(flag=FLAGS.INLINE_POLICIES, key='inline_policies')
+def get_inline_policies(role, **conn):
+    return get_role_inline_policies(role, **conn)
+
+
+@FlagRegistry.register(flag=FLAGS.INSTANCE_PROFILES, key='instance_profiles')
+def get_instance_profiles(role, **conn):
+    return get_role_instance_profiles(role, **conn)
+
+
+def get_role(role, output='camelized', flags=FLAGS.ALL, **conn):
     """
     Orchestrates all the calls required to fully build out an IAM Role in the following format:
 
@@ -56,12 +85,6 @@ def get_role(role, output='camelized', **conn):
     role = modify(role, 'camelized')
     _conn_from_args(role, conn)
     role = _get_base(role, **conn)
-    role.update(
-        {
-            'managed_policies': get_role_managed_policies(role, **conn),
-            'inline_policies': get_role_inline_policies(role, **conn),
-            'instance_profiles': get_role_instance_profiles(role, **conn),
-            '_version': 1
-        }
-    )
+
+    FlagRegistry.build_out(role, flags, role, **conn)
     return modify(role, format=output)
